feat: move infraction to bot
