+++ packages/bot/src/lib/connect-mysql.ts
@@ -4,9 +4,8 @@
 
 import { createConnection, Connection } from 'typeorm'
 import { HelpChannel } from '../entities/help-channel'
-// import { Infractions } from '../entities/infractions'
+import { Infractions } from '../entities/infractions'
 import { MessageRoles, MessageRolesActions } from '../entities/roles'
-// import { AccessToken } from '../entities/access-token'
 
 export const connectMySQL = (): Promise<Connection> => {
   return process.env.NODE_ENV === 'development'
@@ -39,13 +38,7 @@ export const connectMySQL = (): Promise<Connection> => {
         // It's bad but we don't have here any strong data in db
         synchronize: true,
         logging: true,
-        entities: [
-          // AccessToken,
-          // Infractions,
-          HelpChannel,
-          MessageRoles,
-          MessageRolesActions,
-        ],
+        entities: [Infractions, HelpChannel, MessageRoles, MessageRolesActions],
         charset: 'utf8mb4_unicode_ci',
         cache: {
           type: 'ioredis',
