+++ packages/server/src/api/index.ts
@@ -3,7 +3,6 @@
 // found in the LICENSE file.
 
 import * as express from 'express'
-import { requiresBotAuth } from '../lib/requires-auth'
 
 const router = express.Router()
 
@@ -11,21 +10,12 @@ const router = express.Router()
 
 // Meta
 import { hello } from './meta'
-
-// Help Channels
-
-// Infractions
-import { createInfraction } from './infractions'
 //#endregion
 
 //#region Usage
 
 // Meta
 router.get('/', hello)
-
-// Infractions
-router.post('/infractions', requiresBotAuth, createInfraction)
-
 //#endregion
 
 export const apiRoutes = router as express.Router
