+++ packages/server/src/lib/stores/index.ts
@@ -3,4 +3,3 @@
 // found in the LICENSE file.
 
 export * from './access-token'
-export * from './infractions'
